+++ dart/lib/src/transport/http_transport.dart
@@ -7,6 +7,7 @@ import '../protocol.dart';
 import '../sentry_options.dart';
 import '../utils.dart';
 import 'noop_encode.dart' if (dart.library.io) 'encode.dart';
+import 'noop_origin.dart' if (dart.library.html) 'origin.dart';
 import 'transport.dart';
 
 /// A transport is in charge of sending the event to the Sentry server.
@@ -15,16 +16,12 @@ class HttpTransport implements Transport {
 
   final Dsn _dsn;
 
-  /// Use for browser stacktrace
-  final String _origin;
-
   CredentialBuilder _credentialBuilder;
 
   final Map<String, String> _headers;
 
-  HttpTransport({@required SentryOptions options, String origin})
+  HttpTransport({@required SentryOptions options})
       : _options = options,
-        _origin = origin,
         _dsn = Dsn.parse(options.dsn),
         _headers = _buildHeaders(sdkIdentifier: options.sdk.identifier) {
     _credentialBuilder = CredentialBuilder(
@@ -36,7 +33,7 @@ class HttpTransport implements Transport {
 
   @override
   Future<SentryId> send(SentryEvent event) async {
-    final data = event.toJson(origin: _origin);
+    final data = event.toJson(origin: eventOrigin);
 
     final body = _bodyEncoder(
       data,