+++ dart/lib/src/transport/origin.dart
@@ -1,0 +1,3 @@
+import 'dart:html';
+
+String get eventOrigin => '${window.location.origin}/';