+++ dart/test/sentry_browser_test.dart
@@ -2,19 +2,10 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 @TestOn('browser')
-import 'package:sentry/browser.dart';
-import 'package:sentry/sentry.dart';
 import 'package:test/test.dart';
 
 import 'test_utils.dart';
 
 void main() {
-  group('SentryBrowserClient', () {
-    test('SentryClient constructor build browser client', () {
-      final client = SentryClient(SentryOptions(dsn: testDsn));
-      expect(client is SentryBrowserClient, isTrue);
-    });
-
-    runTest(isWeb: true);
-  });
+  runTest(isWeb: true);
 }