+++ dart/test/sentry_io_client_test.dart
@@ -4,19 +4,10 @@
 @TestOn('vm')
 import 'dart:io';
 
-import 'package:sentry/sentry.dart';
-import 'package:sentry/src/sentry_io_client.dart';
 import 'package:test/test.dart';
 
 import 'test_utils.dart';
 
 void main() {
-  group(SentryIOClient, () {
-    test('SentryClient constructor build io client', () {
-      final client = SentryClient(SentryOptions(dsn: testDsn));
-      expect(client is SentryIOClient, isTrue);
-    });
-
-    runTest(gzip: gzip);
-  });
+  runTest(gzip: gzip);
 }